Enforce consistent signed units.

+++ src/Framework/Model.h
@@ -168,7 +168,7 @@ namespace IntegerWorld
 
 		rotation_angle_t rotation;
 		vertex16_t origin;
-		uint32_t radiusSquared;
+		int32_t radiusSquared;
 
 		bool IsPointInside(const vertex16_t& point, const uint16_t planeTolerance = VERTEX16_UNIT / 16) const
 		{
@@ -196,7 +196,7 @@ namespace IntegerWorld
 			const int16_t dx = point.x - origin.x;
 			const int16_t dy = point.y - origin.y;
 			const int16_t dz = point.z - origin.z;
-			const uint32_t squareDistance = (uint32_t(int32_t(dx) * dx) + uint32_t(int32_t(dy) * dy) + uint32_t(int32_t(dz) * dz));
+			const int32_t squareDistance = (int32_t(dx) * dx) + (int32_t(dy) * dy) + (int32_t(dz) * dz);
 
 			// If point is outside the bounding sphere, it's definitely outside the frustum.
 			if (squareDistance > radiusSquared)
